Adds a global `rootRequire` function
This commit adds a `rootRequire` function that wraps the traditional
Node.js `require` function. The point of this is that we can "require"
modules from anywhere in the project relative to the root of the
project without having to specify paths like "../core/some-module.js".

The added benefit is that we can now "require" singleton objects and
they will be properly cached according to their path name by Node.js.